X11: Cache cursor objects in XConnection
